🐛 fix(agent): keep working directory picker when a remote device is selected
Selecting a remote execution device sets runtimeMode to 'none', so the
RuntimeConfig bar's `rightContent()` fell through to `return null` and the
directory picker (and git status) disappeared. The sibling WorkingDirectoryBar
already handled device mode via DeviceWorkingDirectory, but RuntimeConfig was
never updated — two bars, divergent code paths.

Fix: check device mode first in `rightContent()` and render the device-scoped
picker. Extract the shared device/local directory logic into a new
WorkingDirectorySection used by both bars so they can't drift again.
